auth-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
google-auth==2.23.4
requests==2.31.0
httpx==0.25.2
PyJWT==2.8.0